NXDRIVE-1546: Filter out KeyboardInterrupt errors
