+++ docs/changes/4.1.0.md
@@ -70,6 +70,7 @@ Release date: `2019-xx-xx`
 - [NXDRIVE-1510](https://jira.nuxeo.com/browse/NXDRIVE-1510): Unskip `test_move_untrash_file_on_parent_with_no_rights()` since [NXP-25066](https://jira.nuxeo.com/browse/NXP-25066) has been resolved sometime ago
 - [NXDRIVE-1536](https://jira.nuxeo.com/browse/NXDRIVE-1536): Use mock'ed objects instead of manual swap
 - [NXDRIVE-1542](https://jira.nuxeo.com/browse/NXDRIVE-1542): Unlock tests parallelization!
+- [NXDRIVE-1546](https://jira.nuxeo.com/browse/NXDRIVE-1546): Filter out `KeyboardInterrupt` errors
 
 ## Minor Changes
